split test into dev prod and add favicon
